moved loader to outside
