+++ src/components/molecules/Loader.tsx
@@ -1,50 +1,163 @@
-import React, { VFC } from 'react';
 import styled from '@emotion/styled';
+// import { Html, useProgress } from '@react-three/drei';
+import { useEffect, useRef, useState, VFC } from 'react';
 import { color, zIndex } from '../../utils/style';
+import { gsap } from 'gsap';
 
 export const Loader: VFC = () => {
+  // const { progress, item } = useProgress();
+
+  const loaderPath = useRef(null);
+  const loader = useRef(null);
+  const electrode = useRef(null)
+  const counter = useRef<HTMLDivElement|null>(null);
+
+  const [isLoaderOpen, setLoaderOpen] = useState(false);
+  const [isAnimating, setIsAnimating] = useState(false);
+
+  const [hidden, setOcclude] = useState(false)
+
+
+
+  
+
+  const paths = {
+    step1: {
+      unfilled: 'M 0 0 V 0 Q 50 0 100 0 V 0 z',
+      inBetween: {
+        curve: 'M 0 0 V 50 Q 50 100 100 50 V 0 z',
+      },
+      filled: 'M 0 0 V 100 Q 50 100 100 100 V 0 z',
+    }
+  };
+
+  const loaderEnd = () => {
+    if (isAnimating) return;
+    setIsAnimating(true);
+    setLoaderOpen(true);
+
+    if (!loaderPath.current || !loader.current || !electrode.current) return;
+    
+    gsap
+      .timeline({
+        onComplete: () => {
+          setIsAnimating(false);
+          setOcclude(true)
+        },
+      })
+      .set(loaderPath.current, {
+        attr: { d: paths.step1.filled },
+      })
+      .set(loader.current, {
+        visibility: 'visible',
+      })
+      
+      .to(
+        loaderPath.current,
+        {
+          duration: 0.8,
+          delay:3.0,
+          ease: 'power4.in',
+          attr: { d: paths.step1.inBetween.curve },
+        },
+        0,
+      )
+      .fromTo(electrode.current, {
+        opacity:1
+      },{
+        duration: 0.2,
+        ease: 'power1',
+        y:-30,
+        opacity:0
+      })
+      .to(loaderPath.current, {
+        duration: 0.2,
+        ease: 'power1',
+        attr: { d: paths.step1.unfilled },
+      })
+      .set(loader.current, {
+        visibility: 'hidden',
+      })
+  };
+
+  
+  
+  
+  //   return <Html center>{progress} % loaded</Html>
+  useEffect(()=>{
+    loaderEnd()
+  },[])
   return (
-    <LoaderAnim>
-      <LoaderLogo>
-        <LogoInner>
-          <li className="front">
-            <div className="elec__item glass__front"></div>
-            <div className="elec__item mica__front"></div>
-          </li>
-          <li className="sideY">
-            <div className="elec__item glass__side"></div>
-            <ul>
-              <li className="elec__item mica__side"></li>
-              <li className="elec__item plate__side">
-                <svg width="44" height="84">
-                  <path d="M22 0 L1 84 L43 84 Z" fillOpacity="0"></path>
-                </svg>
-              </li>
-            </ul>
-          </li>
-          <li className="sideX">
-            <div className="elec__item glass__side"></div>
-            <ul>
-              <li className="elec__item mica__side"></li>
-              <li className="elec__item plate__side">
-                <svg width="44" height="84">
-                  <path d="M22 0 L1 84 L43 84 Z" fillOpacity="0"></path>
-                </svg>
-              </li>
-            </ul>
-          </li>
-        </LogoInner>
-      </LoaderLogo>
-      <LoaderNum>100</LoaderNum>
-    </LoaderAnim>
+    <Container ref={loader}>
+      <SVG
+        width="100vw"
+        height="100vh"
+        viewBox="0 0 100 100"
+        preserveAspectRatio="none"
+        
+      >
+        <path
+          fill={color.background.middleDark}
+          vectorEffect="non-scaling-stroke"
+          d="M 0 0 V 100 Q 50 100 100 100 V 0 z"
+          ref={loaderPath}
+        />
+      </SVG>
+      <LoaderAnim ref={electrode}>
+        <LoaderLogo>
+          <LogoInner>
+            <li className="front">
+              <div className="elec__item glass__front"></div>
+              <div className="elec__item mica__front"></div>
+            </li>
+            <li className="sideY">
+              <div className="elec__item glass__side"></div>
+              <ul>
+                <li className="elec__item mica__side"></li>
+                <li className="elec__item plate__side">
+                  <svg width="44" height="84">
+                    <path d="M22 0 L1 84 L43 84 Z" fillOpacity="0"></path>
+                  </svg>
+                </li>
+              </ul>
+            </li>
+            <li className="sideX">
+              <div className="elec__item glass__side"></div>
+              <ul>
+                <li className="elec__item mica__side"></li>
+                <li className="elec__item plate__side">
+                  <svg width="44" height="84">
+                    <path d="M22 0 L1 84 L43 84 Z" fillOpacity="0"></path>
+                  </svg>
+                </li>
+              </ul>
+            </li>
+          </LogoInner>
+        </LoaderLogo>
+        {/* <LoaderNum ref={counter} data-from="0" data-to="4096">{progress}</LoaderNum> */}
+
+      </LoaderAnim>
+      
+    </Container>
   );
 };
 
+const Container = styled.div`
+  width:100vw;
+  height:100vh;
+  position:absolute;
+  z-index:${zIndex.elevation.ev16};
+
+`
+
+const SVG=styled.svg`
+    position: absolute;
+    top: 0;
+    left: 0;
+
+`
 const LoaderAnim = styled.div`
-  width: 100%;
-  height: 100vh;
-  position: absolute;
-  background: #000;
+    
   .front {
     animation: rotateFront 5000ms linear infinite;
   }
@@ -137,7 +250,7 @@ const LogoInner = styled.ol`
   transform: translateY(-60px) translateZ(-60px);
 `;
 
-const LoaderNum = styled.p`
+const LoaderNum = styled.div`
   position: absolute;
   top: 50%;
   margin-top: 50px;
